+++ LibOS/shim/src/sys/shim_fork.c
@@ -10,10 +10,10 @@
 #include "shim_internal.h"
 #include "shim_table.h"
 
-int shim_do_fork(void) {
+long shim_do_fork(void) {
     return shim_do_clone(SIGCHLD, 0, NULL, NULL, 0);
 }
 
-int shim_do_vfork(void) {
+long shim_do_vfork(void) {
     return shim_do_clone(CLONE_VFORK | CLONE_VM | SIGCHLD, 0, NULL, NULL, 0);
 }
